Adding multigrid cycles draft and moving all sizes to size_t

+++ includes/src_includes/printing.h
@@ -2,5 +2,5 @@
 #include <vector>
 
 
-void print_matrix(const std::vector<float> &vector, const int num_rows, const int num_cols, const int precision = 3);
+void print_matrix(const std::vector<float> &vector, const size_t num_rows, const size_t num_cols, const int precision = 3);
 void print_vector(const std::vector<float> &vector, const int precision = 3);
