+++ includes/src_includes/restriction.h
@@ -3,4 +3,4 @@
 
 
 void restrict_vector(const std::vector<float> &fine_array, std::vector<float> &coarse_array);
-void restrict_matrix(const std::vector<float> &fine_array, const int fine_rows, const int fine_cols, std::vector<float> &coarse_array, const int coarse_rows, const int coarse_cols);
+void restrict_matrix(const std::vector<float> &fine_array, const size_t fine_rows, const size_t fine_cols, std::vector<float> &coarse_array, const size_t coarse_rows, const size_t coarse_cols);
